mindtrace/automation/pyproject.toml
[project]
name = "mindtrace-automation"
version = "0.1.0"
dependencies = [
    "mindtrace-core",
    "mindtrace-registry",
    "mindtrace-database",
    "mindtrace-services",
    "mindtrace-datalake",
    "mindtrace-models",
    "mindtrace-cluster",
    "mindtrace-storage",
    "tqdm>=4.67.1",
    "pandas>=2.3.1",
    "mtrix",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "huggingface>=0.0.1",
    "opencv-python>=4.11.0.86",
    "reflex>=0.8.2",
    "streamlit>=1.46.1",
    "label-studio-sdk==1.0.17",
    "supervision>=0.23.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
mtrix = { git = "ssh://git@github.com/mindtrace/mt-rix.git", branch="modelbank/feature/v1"}
